+++ Doctrine/TranslationManager.php
@@ -11,10 +11,12 @@
 
 namespace Asm\TranslationLoaderBundle\Doctrine;
 
+use Asm\TranslationLoaderBundle\Event\TranslationEvent;
 use Asm\TranslationLoaderBundle\Model\TranslationInterface;
 use Asm\TranslationLoaderBundle\Model\TranslationManager as BaseTranslationManager;
 use Doctrine\Common\Persistence\ObjectManager;
 use Doctrine\Common\Persistence\ObjectRepository;
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
 /**
  * TranslationManager implementation supporting Doctrine.
@@ -34,12 +36,14 @@ class TranslationManager extends BaseTranslationManager
     private $repository;
 
     /**
-     * @param ObjectManager $objectManager Object manager for translation entities
-     * @param string        $class         Translation model class name
+     * @param ObjectManager            $objectManager   Object manager for translation entities
+     * @param string                   $class           Translation model class name
+     * @param EventDispatcherInterface $eventDispatcher Event dispatcher used to propagate new, modified
+     *                                                  and removed translations
      */
-    public function __construct(ObjectManager $objectManager, $class)
+    public function __construct(ObjectManager $objectManager, $class, EventDispatcherInterface $eventDispatcher)
     {
-        parent::__construct($class);
+        parent::__construct($class, $eventDispatcher);
 
         $this->objectManager = $objectManager;
         $this->repository = $objectManager->getRepository($class);
@@ -68,8 +72,16 @@ class TranslationManager extends BaseTranslationManager
     {
         $translation->setDateUpdated(new \DateTime());
 
+        if($this->objectManager->contains($translation)) {
+            $eventName = TranslationEvent::POST_UPDATE;
+        } else {
+            $eventName = TranslationEvent::POST_PERSIST;
+        }
+
         $this->objectManager->persist($translation);
         $this->objectManager->flush();
+
+        $this->eventDispatcher->dispatch($eventName, new TranslationEvent($translation));
     }
 
     /**
@@ -79,5 +91,7 @@ class TranslationManager extends BaseTranslationManager
     {
         $this->objectManager->remove($translation);
         $this->objectManager->flush();
+
+        $this->eventDispatcher->dispatch(TranslationEvent::POST_REMOVE, new TranslationEvent($translation));
     }
 }