+++ EventListener/TranslationHistorySubscriber.php
@@ -1,106 +1,72 @@
 <?php
-/**
- * @namespace
+
+/*
+ * This file is part of the AsmTranslationLoaderBundle package.
+ *
+ * (c) Marc Aschmann <maschmann@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
  */
+
 namespace Asm\TranslationLoaderBundle\EventListener;
 
-use Doctrine\Common\EventSubscriber;
-use Doctrine\ORM\Event\LifecycleEventArgs;
-use Asm\TranslationLoaderBundle\Entity\Translation;
-use Asm\TranslationLoaderBundle\Entity\TranslationHistory;
+use Asm\TranslationLoaderBundle\Event\TranslationEvent;
+use Asm\TranslationLoaderBundle\Model\TranslationHistoryInterface;
+use Asm\TranslationLoaderBundle\Model\TranslationHistoryManagerInterface;
+use Symfony\Component\Security\Core\SecurityContextInterface;
 
 /**
  * Class TranslationHistorySubscriber
  *
- * @package Asm\TranslationLoaderBundle\EventListener
  * @author marc aschmann <maschmann@gmail.com>
- * @uses Doctrine\Common\EventSubscriber
- * @uses Doctrine\ORM\Event\LifecycleEventArgs
- * @uses Asm\TranslationLoaderBundle\Entity\Translation
- * @uses Asm\TranslationLoaderBundle\Entity\TranslationHistory
+ * @author Christian Flothmann <christian.flothmann@xabbuh.de>
  */
-class TranslationHistorySubscriber implements EventSubscriber
+class TranslationHistorySubscriber
 {
     /**
-     * @var \Symfony\Component\Security\Core\SecurityContext $securityContext
+     * @var TranslationHistoryManagerInterface
+     */
+    private $translationHistoryManager;
+
+    /**
+     * @var SecurityContextInterface
      */
     private $securityContext;
 
-    /**
-     * @return array
-     */
-    public function getSubscribedEvents()
+    public function __construct(TranslationHistoryManagerInterface $translationHistoryManager, SecurityContextInterface $securityContext)
     {
-        return array(
-            'postPersist',
-            'postUpdate',
-            'postRemove',
-        );
-    }
-
-    /**
-     * @param LifecycleEventArgs $args
-     */
-    public function postUpdate(LifecycleEventArgs $args)
-    {
-        $this->index($args, 'update');
-    }
-
-    /**
-     * @param LifecycleEventArgs $args
-     */
-    public function postPersist(LifecycleEventArgs $args)
-    {
-        $this->index($args, 'persist');
-    }
-
-    /**
-     * @param LifecycleEventArgs $args
-     */
-    public function postRemove(LifecycleEventArgs $args)
-    {
-        $this->index($args, 'remove');
-    }
-
-    /**
-     * @param LifecycleEventArgs $args
-     * @param string $type type of event
-     */
-    public function index(LifecycleEventArgs $args, $type)
-    {
-        $entity        = $args->getEntity();
-        $entityManager = $args->getEntityManager();
-
-        if ($entity instanceof Translation) {
-
-            $token    = $this->securityContext->getToken();
-            $userName = 'anonymous';
-            if (!empty($token)) {
-                $userName = $token->getUsername();
-            }
-
-            /** @var \Asm\TranslationLoaderBundle\Entity\TranslationHistory $historyEvent */
-            $historyEvent = new TranslationHistory();
-            $historyEvent->setTransKey($entity->getTransKey());
-            $historyEvent->setTransLocale($entity->getTransLocale());
-            $historyEvent->setMessageDomain($entity->getMessageDomain());
-            $historyEvent->setTranslation($entity->getTranslation());
-            $historyEvent->setUserAction($type);
-            $historyEvent->setUserName($userName);
-            $historyEvent->setDateOfChange(new \DateTime());
-
-            $entityManager->persist($historyEvent);
-            $entityManager->flush();
-            $entityManager->detach($historyEvent);
-            $entityManager->clear();
-        }
-    }
-
-    /**
-     * @param $securityContext
-     */
-    public function setSecurityContext($securityContext)
-    {
+        $this->translationHistoryManager = $translationHistoryManager;
         $this->securityContext = $securityContext;
+    }
+
+    /**
+     * Event-based update of the translation history.
+     *
+     * @param TranslationEvent $event The event triggering the update
+     *
+     * @return TranslationHistoryInterface The new translation history entry
+     */
+    public function updateHistory(TranslationEvent $event)
+    {
+        $translation = $event->getTranslation();
+
+        $historyEntry = $this->translationHistoryManager->createTranslationHistory();
+        $historyEntry->setDateOfChange(new \DateTime());
+        $historyEntry->setMessageDomain($translation->getMessageDomain());
+        $historyEntry->setTransKey($translation->getTransKey());
+        $historyEntry->setTransLocale($translation->getTransLocale());
+        $historyEntry->setTranslation($translation->getTranslation());
+        $historyEntry->setUserAction(strtolower(substr($event->getName(), 4)));
+
+        if (null !== $token = $this->securityContext->getToken()) {
+            $historyEntry->setUserName($token->getUsername());
+        } else {
+            $historyEntry->setUserName('anonymous');
+        }
+
+        $this->translationHistoryManager->updateTranslationHistory($historyEntry);
+
+        return $historyEntry;
     }
 }